chatbot/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
google-cloud-bigquery==3.25.0
faiss-cpu==1.8.0.post1
sentence-transformers==3.1.1
pydantic==2.9.2
python-dateutil==2.9.0.post0